backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langgraph>=0.0.26
langchain>=0.1.6
langchain-openai>=0.0.5
pandas>=2.2.0
numpy>=1.26.3
openpyxl>=3.1.2
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
matplotlib>=3.8.2
seaborn>=0.13.2
scikit-learn>=1.4.0
python-multipart>=0.0.6
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
